fix missing fields check for mappings to substructures
This happens for rcl_interfaces/Log

'header.stamp' gets mapped to 'stamp'. For the old check on both messages where some fields missing: 'topics' in ros2 and 'stamp' in ros1. That 'stamp' was mapped via the custom rule above wasn't recognized by the old logic. For now we say if there a custom field mappings for a field we skip this check.